fix: allow compatibility with lazy.nvim
By using `packadd`, the plugin manager `lazy.nvim` throws an error
regargind not beind able to find the plugin triying to be loaded. This
happens because it already handles lazy loading.

To solve this issue, if-statements were added to all `packadd` calls
with the condition that they will only be called if `lazy.nvim` is not
loaded.